Add files via upload
This Angular frontend application is part of a MEAN stack solution that provides a user-friendly interface for interacting with data from various collections. The application enables users to search, analyze, and visualize data in an intuitive manner. It also supports advanced features like downloading data as Excel or CSV files, storing user filter selections in memory, and visualizing forecast data through graphs and pie charts. The frontend is designed to be modular, customizable, and optimized for performance, ensuring a seamless user experience.